Set DMX values by making requests to dmx-proxy

+++ service.dmx/universe/universe.go
@@ -9,14 +9,14 @@ import (
 
 // Universe represents a set of fixtures in a 512 channel space
 type Universe struct {
-	Number int
+	Number uint8
 
 	fixtures map[string]domain.Fixture
 	mux      *sync.RWMutex
 }
 
 // New returns an initialised universe
-func New(number int) *Universe {
+func New(number uint8) *Universe {
 	return &Universe{
 		Number:   number,
 		fixtures: make(map[string]domain.Fixture),
